Support package update on BeeGFS clusters
Package update hangs when using j1.small instances for OSS nodes, with
the `yum` process stuck in `balance_dirty_pages.isra.20`. This only
happens when applying updates via Ansible, not when running `yum update`
directly from a shell. It is possible that the extra memory consumption
caused by Ansible/Python triggers this behaviour.

Switch to j2.small instances which have twice the amount of RAM to avoid
this issue.